backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.0.0"
description = "mvp poc backend"
authors = ["redliu312 <tommyandrora@gmail.com>"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12"
django = "4.2.11"
djangorestframework = "3.14.0"
django-cors-headers = "4.3.1"
djangorestframework-simplejwt = "^5.5.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"